requirements.txt
agents==1.4.0
aiohttp==3.12.13
fastapi==0.116.1
openai==1.95.1
Pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
supabase==2.16.0
tenacity==9.1.2
openai-agents==0.1.0